GOTCHA: Div instead of Button
This was put in as a Gotcha and waited to be uploaded. The reason I went with a `div` tag instead of a `button` tag is not because the styling is easier, but because I've gotten dismayed with interviewers not actually reading my code submissions.

There is however ONE `div` element on the footer due to that not being a button, but a spacer division in DOM logic